Create BreezBridge for new SDK and cleanup older SDK (#201)
* Use listLsps from Rust library

* Create LightningToolkit service for binding.rs methods

* Rename LightningToolkit to BreezBridge

Add BreezBridge to Injector

* Remove FiatService

* Remove LSPService & integrate LspAPI

* Replace breez_sdk.NodeState with NodeState and all affected Int64->int conversions

Replace Invoice with LNInvoice
Use _createSwapInvoice via rust lib

* Remove unused imports

* Remove sdk implementations on Dart layer

* Remove sdk implementations on Dart layer cont.

* Remove references to SDK implementations on Dart layer

* Set BreezBridge of CurrencyBloc to a private field

* Remove residual fixnum and remove breez server methods keeping just
three that needs to be implmented

* Reflect changes in rust library to BreezBridge

Read and create a Config from breez.conf

* Implement methods needed to register & start a registered node

- _watchAccountChanges()
- store and restore credentials

+++ lib/bloc/withdraw/withdraw_funds_state.dart
@@ -1,4 +1,4 @@
-import 'package:fixnum/fixnum.dart';
+
 
 abstract class WithdrawFudsState {
   const WithdrawFudsState();
@@ -35,7 +35,7 @@ class WithdrawFudsInfoState extends WithdrawFudsState {
 
 class TransactionCost {
   final Duration waitingTime;
-  final Int64 fee;
+  final int fee;
 
   const TransactionCost(
     this.waitingTime,
